refactor(server): move http/api shared libs into server/http and server/lib (PR 1/5)
No behavior change. Pre-refactor move: infrastructure files migrate into their
semantic homes, and new router-level middleware is introduced (unused yet — PR 3
will consume them when it deboilerplate handlers).

Moves:
  server/api/lib/{cors,rateLimit,validate,schemas,jsonSafe}.js → server/http/
  server/api/nutrition/lib/anthropicFetch.js → server/lib/anthropic.js
  server/api/nutrition/lib/nutritionResponse.js → server/lib/

Splits server/http/middleware.js into requestId.js + requestLog.js
(middleware.js kept as a deprecation shim and will be removed later).

New middleware (added but not wired yet):
  http/asyncHandler.js        (moved `wrap()` out of app.js)
  http/setModule.js           (replaces 25× setRequestModule calls)
  http/requireSession.js      (replaces 14× getSessionUser + 401)
  http/requireAnthropicKey.js (replaces 11× ANTHROPIC_API_KEY gate)
  http/requireApiSecret.js    (for /api/push/send-style endpoints)
  http/requireAiQuota.js      (wraps assertAiQuota for router chain)
  http/requireNutritionToken.js (moved token guard out of nutritionSecurity)
  lib/externalHttp.js         (unifies recordMono/Privat/Send/Lookup)

nutritionSecurity.js kept as a deprecation shim re-exporting from the new
locations — stays valid through PR 3, which will delete it.
