+++ server/aiQuota.js
@@ -1,6 +1,6 @@
 import { getSessionUser } from "./auth.js";
 import pool from "./db.js";
-import { getIp } from "./api/lib/rateLimit.js";
+import { getIp } from "./http/rateLimit.js";
 import { logger } from "./obs/logger.js";
 import { aiQuotaBlocksTotal, aiQuotaFailOpenTotal } from "./obs/metrics.js";
