+++ server/api/weekly-digest.js
@@ -1,10 +1,7 @@
 import { assertAiQuota } from "../aiQuota.js";
-import { setCorsHeaders } from "./lib/cors.js";
+import { setCorsHeaders } from "../http/cors.js";
 import { setRequestModule } from "../obs/requestContext.js";
-import {
-  anthropicMessages,
-  extractAnthropicText,
-} from "./nutrition/lib/anthropicFetch.js";
+import { anthropicMessages, extractAnthropicText } from "../lib/anthropic.js";
 
 function extractJsonObject(raw) {
   if (typeof raw !== "string") return null;
